display list of entities in the drawer
